Refactor : Mock API 전역 상태 관리 수정
- Recoil 내에서 Mock API 비동기 처리 및 전역으로 사용
